pyproject.toml
[tool.poetry]
name = "tp3"
version = "0.1.0"
description = "ITBA-TP3"
authors = ["Florencia Riccomi"]

[tool.poetry.dependencies]
python = "^3.8"
pip = "^21.3.1"
SQLAlchemy = "^1.4.26"
psycopg2-binary = "^2.9.1"
pandas = "^1.3.4"
python-decouple = "^3.5"
requests = "^2.26.0"



[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
flake8 = "^4.0.1"
copilot = "^0.1.9"
black = {version = "^21.10b0", allow-prereleases = true}
pdbpp = "^0.10.3"
mypy = "^0.910"
memory-profiler = "^0.58.0"
Sphinx = "^4.3.0"
flake8-docstrings = "^1.6.0"
pytest = "^6.2.5" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
